// linux/northstar/ui/src/gen/northstar_pb.ts
export class Website extends Message<Website> {
  /**
   * @generated from field: uint32 id = 1;
   */
  id = 0;

  /**
   * @generated from field: uint32 service_id = 2;
   */
  serviceId = 0;

  /**
   * @generated from field: string name = 3;
   */
  name = "";

  /**
   * @generated from field: string url = 4;
   */
  url = "";

  /**
   * @generated from field: string username = 5;
   */
  username = "";

  /**
   * @generated from field: optional int32 password_index = 6;
   */
  passwordIndex?: number;

  /**
   * @generated from field: string old_password = 7;
   */
  oldPassword = "";

  /**
   * @generated from field: bool enumerated = 8;
   */
  enumerated = false;

  constructor(data?: PartialMessage<Website>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "northstar.Website";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "id", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 2, name: "service_id", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 3, name: "name", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 4, name: "url", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 5, name: "username", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 6, name: "password_index", kind: "scalar", T: 5 /* ScalarType.INT32 */, opt: true },
    { no: 7, name: "old_password", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 8, name: "enumerated", kind: "scalar", T: 8 /* ScalarType.BOOL */ },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): Website {
    return new Website().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): Website {
    return new Website().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): Website {
    return new Website().fromJsonString(jsonString, options);
  }

  static equals(a: Website | PlainMessage<Website> | undefined, b: Website | PlainMessage<Website> | undefined): boolean {
    return proto3.util.equals(Website, a, b);
  }
}